Migrate MapAccessExprTest
