+++ modules/ballerina-test/src/test/resources/test-src/types/map/map-access-expr.bal
