+++ modules/ballerina-test/src/test/resources/test-src/types/map/nested-map-access.bal
